refactor(ui,core): name the diff row model and review adapter for their roles
Follow-up to the module renames, covering the names that were judgment calls
rather than outright collisions.

- `ui/diff/pierre.ts` was named for the vendor while actually owning `DiffRow`,
  the unit the rest of the UI measures, windows, and paints. It is now
  `diffRows.ts`, and `PierreDiffView` — the body `DiffSection` renders beside
  `FileView` — is `DiffSectionBody`. Pierre is still the highlighter this module
  calls; it is no longer what the module is called.
- `core/loaders.ts` both reached out to the world and built the model. The pure
  half moves to `changesetFromPatch.ts` (patch text in, `DiffFile`s out, no I/O)
  and the rest becomes `changesetSources.ts` — one loader per source shape, each
  ending in that same call.
- `useReviewController` is an adapter, not a controller, and its own header
  already said so; it is now `useTerminalReview`. `ui/lib/reviewProjection.ts`
  shared the word "projection" with the unrelated `projectReviewDocument`, so it
  is now `reviewNoteMapping.ts`, which is all it ever converted.

`AppBootstrap` is deliberately left alone. It looked like a weak name until
`SessionBootstrapResult` and `extensionBootstrap.ts` turned up alongside it:
"bootstrap" is an established convention here for the resolved inputs a
subsystem starts from, and renaming one of the three would break the set.

Renames and one module split: no behavior or signature changes.
