+++ src/core/changesetSources.test.ts
@@ -15,7 +15,10 @@ import { getBundledVcsCatalog } from "../app/vcsCatalog";
 import { createGitVcsAdapter } from "../extensions/default/vcs/git";
 import { toInternalVcsAdapter } from "../extensions/runExtension";
 import { createVcsCatalog } from "./vcs";
-import { loadAppBootstrap as loadCoreAppBootstrap, type LoadAppBootstrapOptions } from "./loaders";
+import {
+  loadAppBootstrap as loadCoreAppBootstrap,
+  type LoadAppBootstrapOptions,
+} from "./changesetSources";
 import type { CliInput } from "./types";
 import type { VcsAdapter } from "./vcs/types";
 import { computeWatchSignature } from "./watch/signature";
