+++ src/ui/AppHost.keybindings.test.tsx
@@ -8,7 +8,7 @@ import { act } from "react";
 import { removeTestDirectory } from "../../test/helpers/filesystem";
 import { resolveConfiguredCliInput } from "../core/config";
 import { getBundledVcsCatalog } from "../app/vcsCatalog";
-import { loadAppBootstrap } from "../core/loaders";
+import { loadAppBootstrap } from "../core/changesetSources";
 import type { AppBootstrap } from "../core/types";
 import { AppHost } from "./AppHost";
