+++ src/ui/components/ui-components.test.tsx
@@ -31,7 +31,7 @@ const { DiffPane } = await import("./panes/DiffPane");
 const { MenuDropdown } = await import("./chrome/MenuDropdown");
 const { StatusBar } = await import("./chrome/StatusBar");
 const { DiffFileHeaderRow } = await import("./panes/DiffFileHeaderRow");
-const { PierreDiffView } = await import("../diff/PierreDiffView");
+const { DiffSectionBody } = await import("../diff/DiffSectionBody");
 const { DiffRowView, measureRenderedRowHeight } = await import("../diff/renderRows");
 
 function createTestDiffFile(
@@ -3355,11 +3355,11 @@ describe("UI components", () => {
     expect(frame).toContain("1 + export const alpha = 2;");
   });
 
-  test("PierreDiffView renders stack-mode wrapped continuation rows", async () => {
+  test("DiffSectionBody renders stack-mode wrapped continuation rows", async () => {
     const file = createWrapBootstrap().changeset.files[0]!;
     const theme = resolveTheme("github-dark-default", null);
     const frame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="stack"
         theme={theme}
@@ -3386,12 +3386,12 @@ describe("UI components", () => {
     expect(addedLines.slice(1).some((line) => line.includes("age';"))).toBe(true);
   });
 
-  test("PierreDiffView can reveal offscreen code columns in nowrap mode", async () => {
+  test("DiffSectionBody can reveal offscreen code columns in nowrap mode", async () => {
     const file = createWrapBootstrap().changeset.files[0]!;
     const theme = resolveTheme("github-dark-default", null);
 
     const baseFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="stack"
         theme={theme}
@@ -3404,7 +3404,7 @@ describe("UI components", () => {
       12,
     );
     const shiftedFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="stack"
         theme={theme}
@@ -3430,7 +3430,7 @@ describe("UI components", () => {
     const width = 64;
 
     const splitFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3443,7 +3443,7 @@ describe("UI components", () => {
       18,
     );
     const stackFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="stack"
         theme={theme}
@@ -3464,7 +3464,7 @@ describe("UI components", () => {
     expect(splitContinuationRows.length).toBeGreaterThan(stackContinuationRows.length);
   });
 
-  test("PierreDiffView anchors range-less notes to the first visible row when hunk headers are hidden", async () => {
+  test("DiffSectionBody anchors range-less notes to the first visible row when hunk headers are hidden", async () => {
     const file = createTestDiffFile(
       "note-fallback",
       "note-fallback.ts",
@@ -3473,7 +3473,7 @@ describe("UI components", () => {
     );
     const theme = resolveTheme("github-dark-default", null);
     const frame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3505,11 +3505,11 @@ describe("UI components", () => {
     );
   });
 
-  test("PierreDiffView shows contextual messages when there is no selected file or no textual hunks", async () => {
+  test("DiffSectionBody shows contextual messages when there is no selected file or no textual hunks", async () => {
     const theme = resolveTheme("github-dark-default", null);
 
     const noFileFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={undefined}
         layout="split"
         theme={theme}
@@ -3523,7 +3523,7 @@ describe("UI components", () => {
     expect(noFileFrame).toContain("No file selected.");
 
     const renameOnlyFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={createEmptyDiffFile("rename-pure")}
         layout="split"
         theme={theme}
@@ -3537,7 +3537,7 @@ describe("UI components", () => {
     expect(renameOnlyFrame).toContain("This change only renames the file.");
 
     const newFileFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={createEmptyDiffFile("new")}
         layout="split"
         theme={theme}
@@ -3551,7 +3551,7 @@ describe("UI components", () => {
     expect(newFileFrame).toContain("The file is marked as new.");
 
     const deletedFileFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={createEmptyDiffFile("deleted")}
         layout="split"
         theme={theme}
@@ -3565,7 +3565,7 @@ describe("UI components", () => {
     expect(deletedFileFrame).toContain("The file is marked as deleted.");
 
     const binaryFileFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={{
           ...createEmptyDiffFile("change"),
           id: "empty:binary",
@@ -3584,12 +3584,12 @@ describe("UI components", () => {
     expect(binaryFileFrame).toContain("Binary file skipped");
   });
 
-  test("PierreDiffView shows the expand chevron only when a source fetcher is attached", async () => {
+  test("DiffSectionBody shows the expand chevron only when a source fetcher is attached", async () => {
     const { file: baseFile } = createExpandableContextDiffFile("expand-affordance", "expand.ts");
     const theme = resolveTheme("github-dark-default", null);
 
     const noFetcherFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={baseFile}
         layout="split"
         theme={theme}
@@ -3609,7 +3609,7 @@ describe("UI components", () => {
     };
 
     const expandableFrame = await captureFrame(
-      <PierreDiffView
+      <DiffSectionBody
         file={fileWithFetcher}
         layout="split"
         theme={theme}
@@ -3624,7 +3624,7 @@ describe("UI components", () => {
     expect(expandableFrame).toContain("▾");
   });
 
-  test("PierreDiffView hides add-note affordances on collapsed and hunk-header rows", async () => {
+  test("DiffSectionBody hides add-note affordances on collapsed and hunk-header rows", async () => {
     const expandable = createExpandableContextDiffFile("meta-hover", "meta-hover.ts");
     const file = {
       ...expandable.file,
@@ -3632,7 +3632,7 @@ describe("UI components", () => {
     };
     const theme = resolveTheme("github-dark-default", null);
     const setup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3695,7 +3695,7 @@ describe("UI components", () => {
     }
   });
 
-  test("PierreDiffView toggles a collapsed gap when clicked", async () => {
+  test("DiffSectionBody toggles a collapsed gap when clicked", async () => {
     const expandable = createExpandableContextDiffFile("expand-click", "expand-click.ts");
     const file = {
       ...expandable.file,
@@ -3704,7 +3704,7 @@ describe("UI components", () => {
     const toggledGaps: string[] = [];
     const theme = resolveTheme("github-dark-default", null);
     const setup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3749,7 +3749,7 @@ describe("UI components", () => {
     }
   });
 
-  test("PierreDiffView highlights expanded unchanged source rows", async () => {
+  test("DiffSectionBody highlights expanded unchanged source rows", async () => {
     const beforeLines = Array.from({ length: 30 }, (_, index) =>
       index === 0
         ? "export const expandedMarker = 1;"
@@ -3767,7 +3767,7 @@ describe("UI components", () => {
     });
     const theme = resolveTheme("github-dark-default", null);
     const setup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3804,7 +3804,7 @@ describe("UI components", () => {
     }
   });
 
-  test("PierreDiffView renders word-diff spans with a visibly different background in split view", async () => {
+  test("DiffSectionBody renders word-diff spans with a visibly different background in split view", async () => {
     const file = createTestDiffFile(
       "word-diff",
       "word-diff.ts",
@@ -3813,7 +3813,7 @@ describe("UI components", () => {
     );
     const theme = resolveTheme("github-dark-default", null);
     const setup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3859,7 +3859,7 @@ describe("UI components", () => {
     }
   });
 
-  test("PierreDiffView reuses highlighted rows after unmounting and remounting a file section", async () => {
+  test("DiffSectionBody reuses highlighted rows after unmounting and remounting a file section", async () => {
     const file = createTestDiffFile(
       "cache",
       "cache.ts",
@@ -3869,7 +3869,7 @@ describe("UI components", () => {
     const theme = resolveTheme("github-dark-default", null);
 
     const firstSetup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3904,7 +3904,7 @@ describe("UI components", () => {
     }
 
     const secondSetup = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={file}
         layout="split"
         theme={theme}
@@ -3943,7 +3943,7 @@ describe("UI components", () => {
       { width: 100, height: 10 },
     );
     const thirdFileCheck = await testRender(
-      <PierreDiffView
+      <DiffSectionBody
         file={files[2]}
         layout="split"
         theme={theme}
